Some cleanups
Tidy up the main readme.
Cleanup the configs and move to a dedicated folder.
Moved all scripts to a dedicated folder.
Removed all unused and empty files.

Removed the need for many gitignore files and the rest gitignore entries
are specific to a local env setup(vscode, liteIde etc) so can be added
to `.git/info/exclude` for each developer.

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>