Added profitability calcs and added profitThreshold in config
Each transaction cost is saved in the db and at the next transaction for
the same slot if the profit is below profitThreshold the transaction is
not executed.

Profit calculations are the same as in the tellor contract.

also moved the GWEI and WEI constants in the common package so that any
package can use them without circular dependencies.

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>