Use a custom tool to generate the golang bindings from etherscan
The tool downloads the etherscan code and replicate some of the abigen
code to generate the bindings.

This simplifies the code quite a bit and makes it deterministic for the
e2e tests since the code is exactly the same one as the one deployed on
rinkeby.

can run it with `make generate-bindings`

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>