Complete revamp of the config handling #326 (#403)
remove GetConfig() and make ParseConfig return a config instance that will be passed to all components that need it.
use strict Json parsing to fail on invalid fields (use the Parse func below)
review all fields and if it is only used by mine move it in the mine config(e.g. ProfitThreshold,Heartbeat,MiningInterruptCheckInterval), if it is only used in dataserver move it in the dataserver config.
make Trackers config a struct and move all fields that are for the trackers in that struct. This will help organize the config a bit better.(e.g. FetchTimeout,MinConfidence,TrackerSleepCycle(rename to SleepCycle),)
rename indexes.json to api.json
remove configFolder and use apiFile, manualDataFile,
remove all unused fields from type Config struct {

Co-authored-by: Krasimir Georgiev <8903888+krasi-georgiev@users.noreply.github.com>