Varios fixes for the new contracts
Updated the default config to not start with a remote db

Enabled admin endpoint for Prometheus to be able to delete old metrics
https://www.shellhacks.com/prometheus-delete-time-series-metrics/

Updated Manual data for USPCE

Added an example in the setup docs how to delete an instance

Added a global validation that all public addresses should start with 0x
and convert all to lower case to be consistent everywhere.

Removed the timeout/lastSubmit tracker as it causes a race condition -
check when the miner last submitted -> the tracker has the old value as
it hasn't updated yet -> the submitter submits although the 15threshold
hasn't passed yet.

Made all configs consistent starting with a capital letter.

Updated the contracts to the latest version.

Improved the logging in various places.

Last submit threshold wasn't working when the miner has never submitted
as it was returning 0 and the time diff calc didn't work.

Changed all hash vars to the new ones. timeOfLastNewValue ->
_TIME_OF_LAST_NEW_VALUE .. etc.

Fixed error handling for failed TX submits

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>